fix: P0 batch 1 — a truncated selector, a replay cache that ignored the MAC, and a gate that could not fail
Three findings from docs/P0-07SEP26.md, each independent and host-testable.

P0-H1: reflex_vm_loader_syscall_valid took int16_t while imm is int32_t, so
the TSYS selector was validated at 16 bits and dispatched at 17. 0x10003
truncated to 3 (SYSCALL_DELAY), passed the loader, and reached the interpreter
as 65539 — where the default handler turns it into a run-time fault. The same
class already fixed for branch targets, two checks away and left behind. Both
truncation directions pinned; the tests fail against the old signature.

The audit's second claim for H1 does not hold and is not fixed: wire selectors
>= 4 are not "wrongly rejected", because no such selector exists. The enum
stops at DELAY = 3, and rejecting an unknown selector is correct.

P0-M1: replay_slot_index XOR'd the trailing MAC bytes in at bit positions 8-23
and then masked the low 6 bits, so the slot was exactly `nonce & 63` for every
sender — the MAC was discarded outright. Measured: 64 peers sharing one nonce
landed in 1 slot. Nonces come from a monotonic microsecond clock, so a peer
whose uptime leads by a multiple of the slot count, or any key-holder choosing
a nonce, evicts a victim's entry and reopens the 5-second window. SECURITY.md
§4 described the blend as working; it never did, and now says so.

The hash moved to goose_policy.c, whose stated purpose is that firmware and
host suite compile the same function rather than a test-side restatement of
the spec — which is exactly the coverage this needed and did not have.

P0-M2: format-check piped clang-format into `head`, and a pipeline's status is
its last command's, so the exit code was discarded; the "passed" echo ran
unconditionally. The gate had been green by construction since it was written,
and 98 of 127 hand-written files had drifted behind it.

Rather than a mass reformat, the gate now checks changed *lines* against the
merge-base: a one-line fix in a legacy file must not demand reformatting the
whole file as the price of the edit. make format-check-all reports the
remaining debt so it stays measurable. clang-format is pinned in CI, since an
unpinned formatter lets the same commit pass today and fail tomorrow.

vm/.clang-format records that subtree's function-brace convention — all 76 of
its function definitions already followed it and the root config contradicted
them, so every edit there had to choose between matching its file and passing
the formatter. core/ and storage/ are Allman the same way and want the same
file when next touched.

Host suite 361/361 (up from 350: 3 loader, 8 replay). Format gate verified to
fail on an introduced violation and pass when clean. Firmware not built — no
hardware attached this pass.

+++ vm/loader.c
@@ -20,7 +20,17 @@ static bool reflex_vm_loader_register_valid(uint8_t index)
     return index < REFLEX_VM_REGISTER_COUNT;
 }
 
-static bool reflex_vm_loader_syscall_valid(int16_t imm)
+/* Takes the immediate at its real width, for the same reason as
+ * reflex_vm_loader_target_in_range below.
+ *
+ * This declared `int16_t imm` while reflex_vm_instruction_t::imm is int32_t,
+ * so the selector was validated at 16 bits and dispatched at 17. A selector
+ * of 0x10003 truncates to 3 (REFLEX_VM_SYSCALL_DELAY), passes here, and then
+ * reaches the interpreter as 65539 — which the default handler rejects as
+ * REFLEX_ERR_NOT_SUPPORTED and turns into a run-time fault. A hole in
+ * defence-in-depth rather than a live escape, but the loader exists to reject
+ * a bad image up front instead of relying on the layer below. */
+static bool reflex_vm_loader_syscall_valid(int32_t imm)
 {
     return imm >= REFLEX_VM_SYSCALL_LOG && imm <= REFLEX_VM_SYSCALL_DELAY;
 }
